+++ executor/executor_bsd.h
@@ -31,7 +31,7 @@ static void os_init(int argc, char** argv, void* data, size_t data_size)
 	setrlimit(RLIMIT_NOFILE, &rlim);
 }
 
-static long execute_syscall(const call_t* c, long a[kMaxArgs])
+static intptr_t execute_syscall(const call_t* c, intptr_t a[kMaxArgs])
 {
 	if (c->call)
 		return c->call(a[0], a[1], a[2], a[3], a[4], a[5], a[6], a[7], a[8]);
